use classes, move mutations
